requirements.txt
numpy==1.21.6
pandas==1.3.5
matplotlib==3.5.2
scikit-learn==1.0.2
joblib==1.1.0
xgboost==1.6.1
lightgbm==4.0.0
tensorflow-cpu==2.10.1
ccxt==2.9.4
streamlit==1.25.0
streamlit-autorefresh==1.0.1
plotly==5.14.1
protobuf==3.20.0
